pdf-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pikepdf==9.0.0
httpx==0.27.0
weasyprint==63.1
pdfminer.six==20231228
python-multipart==0.0.9
jinja2==3.1.4
